projectt/violence-detection-system-main/requirements.txt
torch
torchvision
opencv-python
numpy
tqdm
Pillow
scikit-learn
